Move 'style' function to utils
